% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmdtext.R
\name{render_rmd_MQ_summary}
\alias{render_rmd_MQ_summary}
\title{wrapper function of rmarkdown::render for opencpu to render a report for maxquant summary}
\usage{
render_rmd_MQ_summary(data_table)
}
\arguments{
\item{data_table}{the table is a matrix/data.frame/tble on the server ond, json formted on the front ond opencpu will do the conversion automatically}
}
\value{
no direct return, but write an output.html to the temp session on the opencpu server
}
\description{
This function works on the opencpu server end, to produce an html file to send to the front end to display.
It also works on the stand alone mode, to generate a report file, output.html in the currrent folder, using a public accessible rmd file on github. Of course, this need internect connection to run
}
\examples{
  render_rmd_MQ_summary()
}
\seealso{
\code{\link{render}}  \code{\link{knit}}
}
