% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmdtext.R
\name{render_rmd_server}
\alias{render_rmd_server}
\title{wrapper function of rmarkdown::render for opencpu}
\usage{
render_rmd_server(rmd_file)
}
\arguments{
\item{rmd_file}{For render_rmd_server, a string of the file name on the same opencpu server in the same package, only for tests}
}
\value{
no direct return, but write an output.html to the temp session on the opencpu server
}
\description{
Three general wrapper function to work with opencpu, which can render source r markdown code into html files, to be displacyed on the client end
}
\examples{
  path_to_input <- system.file("rmd", rmd_file, package = "rmdocpu")
  render_rmd_server(path_to_input)
}
\seealso{
\code{\link{render}}  \code{\link{knit}}
}
