% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmdtext.R
\name{render_rmd_url}
\alias{render_rmd_url}
\title{wrapper function of rmarkdown::render for opencpu}
\usage{
render_rmd_url(rmd_url)
}
\arguments{
\item{rmd_url}{For render_rmd_url, a string of public accessible url}
}
\value{
no direct return, but write an output.html to the temp session on the opencpu server
}
\description{
Three general wrapper function to work with opencpu, which can render source r markdown code into html files, to be displacyed on the client end
}
\examples{
  render_rmd_url("https://raw.githubusercontent.com/ningzhibin/rmdocpu/master/inst/rmd/input.Rmd")
}
\seealso{
\code{\link{render}}  \code{\link{knit}}
}
