% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmdtext.R
\name{rmdtext}
\alias{rmdtext}
\title{wrapper function of rmarkdown::render for opencpu}
\usage{
rmdtext(text)
}
\arguments{
\item{text}{For rmdtext, a string, which is the rmardown coding itself}
}
\value{
no direct return, but write an output.html to the temp session on the opencpu server
}
\description{
Three general wrapper function to work with opencpu, which can render source r markdown code into html files, to be displacyed on the client end
}
\examples{
  path_to_input <- system.file("rmd", rmd_file, package = "rmdocpu")
  myfile <- readLines(path_to_input)
  rmdtext(myfile)
}
\seealso{
\code{\link{render}}  \code{\link{knit}}
}
